solve day 11

+++ day11.hpp
@@ -8,7 +8,7 @@ namespace Day11 {
 using namespace std;
 
 long solve_day11_pt1(const string &input);
-int solve_day11_pt2(const string &input);
+long solve_day11_pt2(const string &input);
 
 } // namespace Day11
